backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.4
pydantic-settings==2.8.1
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.5.0
prophet==1.1.5
torch==2.2.2
google-genai==1.13.0
qdrant-client==1.13.2
huggingface-hub==0.30.2
pdfplumber==0.11.5
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
